+++ ext_emconf.php
@@ -1,7 +1,7 @@
 <?php
 $EM_CONF[$_EXTKEY] = [
     'title' => 'Feed Generator',
-    'description' => 'Generator for Atom and RSS feeds',
+    'description' => 'Generator for Atom, JSON and RSS feeds',
     'category' => 'fe',
     'author' => 'Chris Müller',
     'author_email' => 'typo3@krue.ml',